Enable providing package versions at build time
The latest version (5.6.0) is incompatible with ES 7.3.0, and we were unable to find a built image with a more recent version. That led us to cook our own image, using the modified Dockerfile in this PR.

I'm submitting this for your consideration, just in case you find the enhancement useful.