chore(diagnostic): Upgrade the diagnostic image to ubuntu latest LTS
The `postgresql-client` version is not sync anymore with the version
in production for our micro-services.
It results in the following error when trying to describe a table:

```
=> \d users
ERROR:  column c.relhasoids does not exist
LINE 1: ...riggers, c.relrowsecurity, c.relforcerowsecurity, c.relhasoi...
```

This commit upgrades the ubuntu version to get the latest postgresql
version (12.9) and hopefully will fix the issue.